move features to @/features
